Created database.js WIP
