+++ github.js
@@ -7,9 +7,6 @@ function readToken() {
         return data;
     } catch (err) {
         console.error(err);
-        //return message.reply(
-        //    "Your GitHub Personal Access Token could not been read. Please set it again using -github-info."
-        // );
     }
 }
